Layout Guide
Made a rough general layout of the site. There are now 3 columns. The rough plan is to have the left column be for account management. The middle column will be the main users page. The right side will be a friends/subscribed list.

It isn't very pretty yet, but it is getting there.