UI Update
Updated UI to look a bit nicer. Site is comming in nicely. Not everything planned is up and showing. Some stuff will wait till later so we can plan a bit more. For the most part this mostly dealt with UI and looks. Some main formatting of the site took place.