Sign in, Log In pages
Moved Sign in page to AccountCreate Page. Had to mess with css a little to make get the boxes to not go wider than the container. Log in page needs to be added to master page.
There also needs to be a clean up of the site and its folders.