Created some generic roles / Updated the tables.
Included the same drop table script.  Have to see if it removes the tables correctly after its ran, check to make sure they are replaced with empty duplicates.

We'll have to discus different roles the program will need to use. I created a gerneral admin view which will allow viewing of all the data without modifications.  And a User Creation role which allows viewing and inserting of data into the User and Security tables.